+++ CHANGELOG.md
@@ -10,6 +10,7 @@ User-visible changes worth mentioning.
 Add your entry here.
 
 - [#1747] Fix unknown pkce method error when configured
+- [#1744] Allow for expired refresh tokens to be revoked
 
 ## 5.8.0